Move some namespaces on Grid
