+++ src/Bundle/Grid/View/LegacyGridViewFactory.php
@@ -17,9 +17,9 @@ use Sylius\Bundle\ResourceBundle\Context\Option\RequestConfigurationOption;
 use Sylius\Component\Grid\Definition\Grid;
 use Sylius\Component\Grid\Parameters;
 use Sylius\Component\Grid\View\GridView;
-use Sylius\Component\Resource\Grid\View\Factory\GridViewFactoryInterface;
 use Sylius\Resource\Context\Context;
 use Sylius\Resource\Context\Option\MetadataOption;
+use Sylius\Resource\Grid\View\Factory\GridViewFactoryInterface;
 
 final class LegacyGridViewFactory implements GridViewFactoryInterface
 {
