+++ src/Bundle/Resources/config/services/state.xml
@@ -130,7 +130,7 @@
             <tag name="sylius.state_responder" />
         </service>
 
-        <service id="Sylius\Component\Resource\Grid\State\RequestGridProvider">
+        <service id="Sylius\Resource\Grid\State\RequestGridProvider">
             <argument type="service" id="sylius.grid.view_factory.resource" on-invalid="null" />
             <argument type="service" id="sylius.grid.service_grid_provider" on-invalid="null" />
             <tag name="sylius.state_provider" />
