+++ src/Component/src/Grid/State/RequestGridProvider.php
@@ -11,16 +11,16 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Grid\State;
+namespace Sylius\Resource\Grid\State;
 
 use Pagerfanta\Pagerfanta;
 use Sylius\Component\Grid\Parameters;
 use Sylius\Component\Grid\Provider\GridProviderInterface;
-use Sylius\Component\Resource\Grid\View\Factory\GridViewFactoryInterface;
 use Sylius\Component\Resource\Metadata\GridAwareOperationInterface;
 use Sylius\Component\Resource\Metadata\Operation;
 use Sylius\Resource\Context\Context;
 use Sylius\Resource\Context\Option\RequestOption;
+use Sylius\Resource\Grid\View\Factory\GridViewFactoryInterface;
 use Sylius\Resource\State\ProviderInterface;
 
 final class RequestGridProvider implements ProviderInterface
