+++ src/Component/src/Grid/View/Factory/GridViewFactory.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Grid\View\Factory;
+namespace Sylius\Resource\Grid\View\Factory;
 
 use Sylius\Component\Grid\Data\DataProviderInterface;
 use Sylius\Component\Grid\Definition\Grid;
