+++ src/Component/src/Grid/View/Factory/GridViewFactoryInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Grid\View\Factory;
+namespace Sylius\Resource\Grid\View\Factory;
 
 use Sylius\Component\Grid\Definition\Grid;
 use Sylius\Component\Grid\Parameters;
